third_party/WebKit/LayoutTests/http/tests/devtools/sources/inspect-function.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/console-test.js"></script>
<script>
function test() {
  var revealed = 0;
  TestRunner.addSniffer(Sources.SourcesView.prototype, 'showSourceLocation', didReveal, true);
  ConsoleTestRunner.evaluateInConsole('function foo() { }; inspect(foo.bind());inspect(foo);');

  function didReveal(uiSourceCode, lineNumber, columnNumber) {
    TestRunner.addResult('Function was revealed:');
    TestRunner.addResult('lineNumber: ' + lineNumber);
    TestRunner.addResult('columnNumber: ' + columnNumber);
    if (revealed == 0)
      ++revealed;
    else
      TestRunner.completeTest();
  }
}
</script>
</head>
<body onload="runTest()">
    <p>Tests that inspect object action works for function and resolve bound function location.</p>
</body>
</html>
