third_party/WebKit/LayoutTests/http/tests/devtools/sources/inspect-function-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that inspect object action works for function and resolve bound function location.

Function was revealed:
lineNumber: 0
columnNumber: 12
Function was revealed:
lineNumber: 0
columnNumber: 12
